Update Exercise 3.44.cpp

+++ Chapter 3/Exercise 3.44.cpp
@@ -31,9 +31,9 @@ int main()
 	
 	//Using for with subscipt Operator
 	cout << "\n\nUsing for with subscript Operator" << endl;
-	for (int i = 0; i < row; ++i)				
+	for (size_t i = 0; i < row; ++i)				
 	{
-		for (int j = 0; j < coloumn; ++j)		
+		for (size_t j = 0; j < coloumn; ++j)		
 			cout << ia[i][j] << " ";
 
 		cout << endl;
